requirements.txt
Django==5.2.10
djangorestframework==3.14.0
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.1
django-filter==24.3
dj-database-url==2.3.0
drf-spectacular==0.27.2
